create AudioNodeChain class
An object of this class represents a chain of AudioNode objects
that can be attached to an existing AudioNode graph.

Soon, we'll define a method in the SongNode class `getAudioNodeChain`
so that each SongNode can define what AudioNode it represents.

We'll make a new Audio class to connect these together as necessary.

Rename the current Audio class to AudioEnv since it's just storing
constants and the AudioContext object.