song-tree: infer parent durations from children
Instead of setting a duration in SongNode,
calculate the duration when the PlayTree is constructed.

This has less chance for bugs like
44c5eb8073fe6d6cb5ad309152a4dbb9fe505d22

Because we don't calculate the value until it's needed,
there is less chance that something will change after
it's calculated.

Also, clean up the PlayNode constructor and add a test.